The interface and the validation of inputs was changed
The sine generator accepts now int, str and enum for the amplitude value. The local input functions were removed as they are not suitable for a remote system. If a value error is raised the sine generator will not start. There is some discussion whether or not the constants in the amplitudes class should be renamed (sk & cr).